Move logic out of backend.js
