+++ server/lib/backend-adapters/video.js
@@ -1,3 +1,5 @@
+import sliceList from '../helpers/sliceList';
+
 class Playlist {
 	constructor(cache) {
 		this.cache = cache;
@@ -7,7 +9,8 @@ class Playlist {
 		return this.cache.cached(`videos.playlist.${id}`, ttl, () => {
 			return fetch(`http://next-video.ft.com/api/playlist/${encodeURI(id)}?videoFields=id,name,renditions,longDescription,publishedDate,videoStillURL`)
 				.then(res => res.json())
-				.then(json => json.items);
+				.then(json => json.items)
+				.then(topics => sliceList(topics, {from, limit}));
 		});
 	}
 }
