+++ server/lib/backend.js
@@ -1,4 +1,3 @@
-import articleGenres from 'ft-next-article-genre';
 import { logger } from 'ft-next-express';
 
 import Cache from './cache';
@@ -19,33 +18,6 @@ import MockLiveblog from './backend-adapters/mock-liveblog';
 
 import capifyMetadata from './helpers/capifyMetadata';
 
-const sliceList = (items, {from, limit}) => {
-	items = items || [];
-	items = (from ? items.slice(from) : items);
-	items = (limit ? items.slice(0, limit) : items);
-
-	return items;
-};
-
-// internal content filtering logic shared for ContentV1 and ContentV2
-const filterContent = ({from, limit, genres, type}, resolveType) => {
-	return (items = []) => {
-		if (genres && genres.length) {
-			items = items.filter(item => genres.indexOf(articleGenres(capifyMetadata(item.metadata), {requestedProp: 'editorialTone'})) > -1);
-		}
-
-		if (type) {
-			if(type === 'liveblog') {
-				items = items.filter(it => resolveType(it) === 'liveblog');
-			} else {
-				items = items.filter(it => resolveType(it) !== 'liveblog');
-			}
-		}
-
-		return sliceList(items, {from, limit});
-	};
-};
-
 class Backend {
 	constructor(adapters, type) {
 		this.adapters = adapters;
@@ -53,26 +25,11 @@ class Backend {
 	}
 
 	page(uuid, sectionsId, ttl = 50) {
-		return this.adapters.capi.page(uuid, ttl)
-		.then(it => ({
-			id: uuid,
-			title: it.title,
-			sectionId: sectionsId,
-			items: it.slice()
-		}))
-		.catch(e => {
-			logger.error(`Error getting page ${uuid}`, e)
-		});
+		return this.adapters.capi.page(uuid, sectionsId, ttl);
 	}
 
 	byConcept(uuid, title, ttl = 50) {
-		return this.adapters.capi.contentAnnotatedBy(uuid, ttl)
-		.then(ids => ({
-			title: title,
-			conceptId: uuid,
-			sectionId: null,
-			items: ids.slice()
-		}));
+		return this.adapters.capi.contentAnnotatedBy(uuid, title, ttl)
 	}
 
 	search(query, ttl = 50) {
@@ -80,107 +37,58 @@ class Backend {
 	}
 
 	content(uuids, opts) {
-		return this.adapters.capi.content(uuids)
-			.then(filterContent(opts, this.resolveContentType));
+		return this.adapters.capi.content(uuids, opts);
 	}
 
 	popular(url, title, ttl = 50) {
-		return this.adapters.popular.fetch(url, ttl)
-		.then((data) => {
-			return data.mostRead.pages.map(function (page) {
-				const index = page.url.lastIndexOf('/');
-				const id = page.url.substr(index + 1).replace('.html', '');
-				return id;
-			});
-		})
-		.then((ids) => ({
-			id: null,
-			sectionId: null,
-			title: title,
-			items: ids
-		}));
+		return this.adapters.popular.fetch(url, title, ttl);
 	}
 
 	liveblogExtras(uri, {limit}, ttl = 50) {
-		return this.adapters.liveblog.fetch(uri, ttl)
-		.then(json => {
-			const dated = json.filter(it => !!it.data.datemodified);
-			const [first, second] = dated.slice(0, 2);
-
-			// make sure updates are in order from latest to earliest
-			if((first && first.data.datemodified) < (second && second.data.datemodified)) { json.reverse(); }
-
-			// dedupe updates and only keep messages, decide on status
-			let [, updates, status] = json.reduce(([skip, updates, status], event) => {
-				if (event.event === 'end') { return [skip, updates, 'closed']; }
-
-				if (event.event === 'msg' && event.data.mid && !skip[event.data.mid]) {
-					updates.push(event);
-					skip[event.data.mid] = true;
-					status = status || 'inprogress';
-				}
-
-				return [skip, updates, status];
-			}, [{}, [], null]);
-
-			if(limit) { updates = updates.slice(0, limit); }
-
-			status = status || 'comingsoon';
-			return {updates, status};
-		});
+		return this.adapters.liveblog.fetch(uri, {limit}, ttl);
 	}
 
 	fastFT() {
 		return this.adapters.fastFT.fetch();
 	}
 
-	resolveContentType(value) {
-		if (/liveblog|marketslive|liveqa/i.test(value.webUrl)) {
-			return 'liveblog';
-		} else {
-			return 'article';
-		}
-	}
 
 	videos(id, {from, limit}, ttl = 50) {
-		return this.adapters.video.playlist(id, ttl)
-			.then(topics => sliceList(topics, {from, limit}));
+		return this.adapters.video.playlist(id,{from, limit}, ttl);
+
 	}
 
 	list(uuid, ttl = 50) {
-		return this.adapters.capi.list(uuid, ttl)
-			// return 'fake' list, so Collection can resolveType correctly
-			.catch(() => ({ apiUrl: `http://api.ft.com/lists/${uuid}` }));
+		return this.adapters.capi.list(uuid, ttl);
 	}
 
 	popularTopics({from, limit}, ttl = 50) {
-		return this.adapters.popularApi.topics(ttl)
-		.then(topics => sliceList(topics, {from, limit}));
+		return this.adapters.popularApi.topics({from, limit}, ttl)
 	}
 
 	popularArticles(args, ttl = 50) {
-		return this.adapters.popularApi.articles(ttl)
-			.then(articles => sliceList(articles, args));
+		return this.adapters.popularApi.articles(args, ttl)
+
 	}
 
 	popularFromHui(args, ttl = 50) {
 		return this.adapters.hui.content(args, ttl)
-			.then(articles => sliceList(articles, args));
+
 	}
 
 	userSavedContent(args, ttl = -1) {
 		return this.adapters.myft.savedContent(args, ttl)
-			.then(res => res.items);
+
 	}
 
 	userFollowedConcepts(args, ttl = -1) {
 		return this.adapters.myft.followedConcepts(args, ttl)
-			.then(res => res.items);
+
 	}
 
 	userPersonalisedFeed(args, ttl = -1) {
 		return this.adapters.myft.personalisedFeed(args, ttl)
-			.then(res => sliceList(res.results, args));
+
 	}
 }
